bunch of cleanup/fixes

+++ src/meshing.cpp
@@ -60,18 +60,18 @@ const std::vector<glm::vec3> cube_bottom = {
     {1.0f, 0.0f, 0.0f}, {0.0f, 0.0f, 0.0f}, {0.0f, 0.0f, 1.0f}};
 
 const std::vector<glm::vec3> cube_up = {{0.0f, 1.0f, 1.0f}, {1.0f, 1.0f, 1.0f},
-					{1.0f, 1.0f, 0.0f}, {1.0f, 1.0f, 0.0f},
-					{0.0f, 1.0f, 0.0f}, {0.0f, 1.0f, 1.0f}};
+                                        {1.0f, 1.0f, 0.0f}, {1.0f, 1.0f, 0.0f},
+                                        {0.0f, 1.0f, 0.0f}, {0.0f, 1.0f, 1.0f}};
 
 const Texture_lookup textures[4] = {{-1, -1, -1, -1, -1, -1},
-				    {1, 1, 1, 1, 1, 1},
-				    {3, 3, 3, 3, 2, 0},
-				    {18, 18, 18, 18, 18, 18}};
+                                    {1, 1, 1, 1, 1, 1},
+                                    {3, 3, 3, 3, 2, 0},
+                                    {18, 18, 18, 18, 18, 18}};
 
 namespace mesher {
 
 const std::vector<Vertex> getFace(const Block &block, glm::ivec3 pos,
-				  enum BlockSide side, glm::vec3 scale) {
+                                  enum BlockSide side, glm::vec3 scale) {
   std::vector<Vertex> vertices;
   std::vector<glm::vec3> positions;
 
@@ -93,7 +93,7 @@ const std::vector<Vertex> getFace(const Block &block, glm::ivec3 pos,
       break;
     case BlockSide::Bottom:
       positions.insert(positions.begin(), cube_bottom.begin(),
-		       cube_bottom.end());
+                       cube_bottom.end());
       break;
     case BlockSide::Up:
       positions.insert(positions.begin(), cube_up.begin(), cube_up.end());
@@ -101,7 +101,7 @@ const std::vector<Vertex> getFace(const Block &block, glm::ivec3 pos,
     default:
       break;
   }
-  for (int i = 0; i < positions.size(); i++) {
+  for (size_t i = 0; i < positions.size(); i++) {
     positions.at(i) *= scale;
   }
   // positions = get_scaled_cube(scale);
@@ -135,8 +135,8 @@ std::vector<Vertex> get_scaled_cube(Block b, glm::vec3 scale, glm::ivec3 pos) {
 }
 
 glm::ivec3 get_interval(Block *data, glm::ivec3 pos,
-			std::vector<glm::ivec2> interval_dimension[3],
-			Block current_block) {
+                        std::vector<glm::ivec2> interval_dimension[3],
+                        Block current_block) {
   glm::ivec3 save_pos = pos;
   glm::ivec3 size = glm::ivec3(0);
 
@@ -165,9 +165,9 @@ glm::ivec3 get_interval(Block *data, glm::ivec3 pos,
   while (pos.z < CHUNK_SIZE) {
     while (size.y > 0 && pos.y - save_pos.y < size.y) {
       while (size.x > 0 && pos.x - save_pos.x < size.x) {
-	front_block = get_block(data, {pos.x, pos.y, pos.z});
-	if (front_block != current_block) break;
-	pos.x++;
+        front_block = get_block(data, {pos.x, pos.y, pos.z});
+        if (front_block != current_block) break;
+        pos.x++;
       }
       if (front_block != current_block) break;
       pos.x = save_pos.x;
@@ -189,13 +189,13 @@ glm::ivec3 get_interval(Block *data, glm::ivec3 pos,
 }
 
 bool is_fill(std::vector<glm::ivec2> interval_dimension[3], glm::ivec3 pos) {
-  for (int i = 0; i < interval_dimension[0].size(); ++i) {
+  for (size_t i = 0; i < interval_dimension[0].size(); ++i) {
     if (interval_dimension[0].at(i).x <= pos.x &&
-	interval_dimension[0].at(i).y > pos.x &&
-	interval_dimension[1].at(i).x <= pos.y &&
-	interval_dimension[1].at(i).y > pos.y &&
-	interval_dimension[2].at(i).x <= pos.z &&
-	interval_dimension[2].at(i).y > pos.z)
+        interval_dimension[0].at(i).y > pos.x &&
+        interval_dimension[1].at(i).x <= pos.y &&
+        interval_dimension[1].at(i).y > pos.y &&
+        interval_dimension[2].at(i).x <= pos.z &&
+        interval_dimension[2].at(i).y > pos.z)
       return true;
   }
   return false;
@@ -204,41 +204,41 @@ bool is_fill(std::vector<glm::ivec2> interval_dimension[3], glm::ivec3 pos) {
 void greedy(Block *data, bool *dirty, RenderAttrib &render_attrib) {
   size_t total_vertices = 0;
   enum BlockSide sides[4] = {BlockSide::Left, BlockSide::Right,
-			     BlockSide::Bottom, BlockSide::Up};
+                             BlockSide::Bottom, BlockSide::Up};
   glm::ivec3 inter = glm::ivec3(0);
   std::vector<glm::ivec2> interval_dimension[3] = {{}};
-  for (int model_id = 0; model_id < MODEL_PER_CHUNK; model_id++) {
+  for (unsigned int model_id = 0; model_id < MODEL_PER_CHUNK; model_id++) {
     if (dirty[model_id] == false) continue;
     std::vector<Vertex> vertices;
     for (int y = model_id * MODEL_HEIGHT; y < ((model_id + 1) * MODEL_HEIGHT);
-	 y++) {
+         y++) {
       for (int x = 0; x < CHUNK_SIZE; x++) {
-	Block current_block = {};
-	for (int z = 0; z < CHUNK_SIZE; z++) {
-	  Block front_block = get_block(data, {x, y, z});
-	  if (front_block.material != Material::Air &&
-	      !is_fill(interval_dimension, {x, y, z})) {
-	    Block b = front_block.material != Material::Air ? front_block
-							    : current_block;
-	    inter =
-		get_interval(data, glm::ivec3(x, y, z), interval_dimension, b);
-	    interval_dimension[0].push_back({x, x + inter.x});
-	    interval_dimension[1].push_back({y, y + inter.y});
-	    interval_dimension[2].push_back({z, z + inter.z});
-	    auto cube = get_scaled_cube(b, inter, {x, y, z});
-	    vertices.insert(vertices.begin(), cube.begin(), cube.end());
-	    current_block = front_block;
-	  }
-	}
+        Block current_block = {};
+        for (int z = 0; z < CHUNK_SIZE; z++) {
+          Block front_block = get_block(data, {x, y, z});
+          if (front_block.material != Material::Air &&
+              !is_fill(interval_dimension, {x, y, z})) {
+            Block b = front_block.material != Material::Air ? front_block
+                                                            : current_block;
+            inter =
+                get_interval(data, glm::ivec3(x, y, z), interval_dimension, b);
+            interval_dimension[0].push_back({x, x + inter.x});
+            interval_dimension[1].push_back({y, y + inter.y});
+            interval_dimension[2].push_back({z, z + inter.z});
+            auto cube = get_scaled_cube(b, inter, {x, y, z});
+            vertices.insert(vertices.begin(), cube.begin(), cube.end());
+            current_block = front_block;
+          }
+        }
       }
     }
     dirty[model_id] = false;
     total_vertices += vertices.size();
     if (vertices.size() > 0) {
       if (render_attrib.vaos.size() <= model_id) {
-	render_attrib.vaos.push_back(new VAO(vertices));
+        render_attrib.vaos.push_back(new VAO(vertices));
       } else {
-	render_attrib.vaos[model_id]->update(vertices);
+        render_attrib.vaos[model_id]->update(vertices);
       }
     }
   }
@@ -247,85 +247,86 @@ void greedy(Block *data, bool *dirty, RenderAttrib &render_attrib) {
 void culling(Block *data, bool *dirty, RenderAttrib &render_attrib) {
   size_t total_vertices = 0;
   enum BlockSide sides[4] = {BlockSide::Left, BlockSide::Right,
-			     BlockSide::Bottom, BlockSide::Up};
-  for (int model_id = 0; model_id < CHUNK_HEIGHT / MODEL_HEIGHT; model_id++) {
+                             BlockSide::Bottom, BlockSide::Up};
+  for (unsigned int model_id = 0; model_id < CHUNK_HEIGHT / MODEL_HEIGHT;
+       model_id++) {
     if (dirty[model_id] == false) continue;
     std::vector<Vertex> vertices;
     for (int y = model_id * MODEL_HEIGHT; y < ((model_id + 1) * MODEL_HEIGHT);
-	 y++) {
+         y++) {
       for (int x = 0; x < CHUNK_SIZE; x++) {
-	Block current_block = {};
-	for (int z = 0; z < CHUNK_SIZE; z++) {
-	  Block front_block = get_block(data, {x, y, z});
-	  if (front_block != current_block) {
-	    Block b = front_block.material != Material::Air ? front_block
-							    : current_block;
-	    auto quad =
-		getFace(b, {x, y, z}, BlockSide::Front, glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	    current_block = front_block;
-	  }
-	  if (z == CHUNK_SIZE - 1 && current_block.material != Material::Air) {
-	    auto quad = getFace(current_block, {x, y, z}, BlockSide::Back,
-				glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	  }
-	  if (x == 0 && current_block.material != Material::Air) {
-	    auto quad = getFace(current_block, {x, y, z}, BlockSide::Right,
-				glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	  }
-	  if (x == CHUNK_SIZE - 1 && current_block.material != Material::Air) {
-	    auto quad = getFace(current_block, {x, y, z}, BlockSide::Left,
-				glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	  }
-	  if (y == 0 && current_block.material != Material::Air) {
-	    auto quad = getFace(current_block, {x, y, z}, BlockSide::Bottom,
-				glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	  }
-	  if (y == CHUNK_HEIGHT - 1 &&
-	      current_block.material != Material::Air) {
-	    auto quad = getFace(current_block, {x, y, z}, BlockSide::Up,
-				glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	  }
-	  if (y == ((model_id + 1) * MODEL_HEIGHT) - 1 &&
-	      current_block.material != Material::Air) {
-	    auto quad = getFace(current_block, {x, y, z}, BlockSide::Up,
-				glm::vec3(1.0f));
-	    vertices.insert(vertices.end(), quad.begin(), quad.end());
-	  }
-	  if (current_block.material == Material::Air) {
-	    glm::ivec3 positions[4] = {
-		glm::ivec3(x - 1, y, z), glm::ivec3(x + 1, y, z),
-		glm::ivec3(x, y + 1, z), glm::ivec3(x, y - 1, z)};
-	    for (int f = 0; f < 4; f++) {
-	      Block b = get_block(data, positions[f]);
-	      if (b.material != Material::Air) {
-		auto quad = getFace(b, positions[f], sides[f], glm::vec3(1.0f));
-		vertices.insert(vertices.end(), quad.begin(), quad.end());
-	      }
-	    }
-	  }
-	}
+        Block current_block = {};
+        for (int z = 0; z < CHUNK_SIZE; z++) {
+          Block front_block = get_block(data, {x, y, z});
+          if (front_block != current_block) {
+            Block b = front_block.material != Material::Air ? front_block
+                                                            : current_block;
+            auto quad =
+                getFace(b, {x, y, z}, BlockSide::Front, glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+            current_block = front_block;
+          }
+          if (z == CHUNK_SIZE - 1 && current_block.material != Material::Air) {
+            auto quad = getFace(current_block, {x, y, z}, BlockSide::Back,
+                                glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+          }
+          if (x == 0 && current_block.material != Material::Air) {
+            auto quad = getFace(current_block, {x, y, z}, BlockSide::Right,
+                                glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+          }
+          if (x == CHUNK_SIZE - 1 && current_block.material != Material::Air) {
+            auto quad = getFace(current_block, {x, y, z}, BlockSide::Left,
+                                glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+          }
+          if (y == 0 && current_block.material != Material::Air) {
+            auto quad = getFace(current_block, {x, y, z}, BlockSide::Bottom,
+                                glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+          }
+          if (y == CHUNK_HEIGHT - 1 &&
+              current_block.material != Material::Air) {
+            auto quad = getFace(current_block, {x, y, z}, BlockSide::Up,
+                                glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+          }
+          if (y == ((model_id + 1) * MODEL_HEIGHT) - 1 &&
+              current_block.material != Material::Air) {
+            auto quad = getFace(current_block, {x, y, z}, BlockSide::Up,
+                                glm::vec3(1.0f));
+            vertices.insert(vertices.end(), quad.begin(), quad.end());
+          }
+          if (current_block.material == Material::Air) {
+            glm::ivec3 positions[4] = {
+                glm::ivec3(x - 1, y, z), glm::ivec3(x + 1, y, z),
+                glm::ivec3(x, y + 1, z), glm::ivec3(x, y - 1, z)};
+            for (int f = 0; f < 4; f++) {
+              Block b = get_block(data, positions[f]);
+              if (b.material != Material::Air) {
+                auto quad = getFace(b, positions[f], sides[f], glm::vec3(1.0f));
+                vertices.insert(vertices.end(), quad.begin(), quad.end());
+              }
+            }
+          }
+        }
       }
     }
     dirty[model_id] = false;
     total_vertices += vertices.size();
     if (vertices.size() > 0) {
       if (render_attrib.vaos.size() <= model_id) {
-	render_attrib.vaos.push_back(new VAO(vertices));
+        render_attrib.vaos.push_back(new VAO(vertices));
       } else {
-	render_attrib.vaos[model_id]->update(vertices);
+        render_attrib.vaos[model_id]->update(vertices);
       }
     }
   }
 }
 
 void get_aabb(Block *data, glm::vec3 &aabb_center, glm::vec3 &aabb_halfsize,
-	      const glm::vec3 chunk_pos) {
+              const glm::vec3 chunk_pos) {
   glm::ivec3 aabb_min;
   glm::ivec3 aabb_max;
   if (aabb_center.x == 0.0f && aabb_center.y == 0.0f && aabb_center.z == 0.0f) {
@@ -340,16 +341,16 @@ void get_aabb(Block *data, glm::vec3 &aabb_center, glm::vec3 &aabb_halfsize,
   for (int y = 0; y < CHUNK_HEIGHT; y++) {
     for (int x = 0; x < CHUNK_SIZE; x++) {
       for (int z = 0; z < CHUNK_SIZE; z++) {
-	Block block = get_block(data, {x, y, z});
-	if (block.material != Material::Air) {
-	  glm::ivec3 pos = {x, y, z};
-	  if (pos.x < aabb_min.x) aabb_min.x = pos.x;
-	  if (pos.x > aabb_max.x) aabb_max.x = pos.x;
-	  if (pos.y < aabb_min.y) aabb_min.y = pos.y;
-	  if (pos.y > aabb_max.y) aabb_max.y = pos.y;
-	  if (pos.z < aabb_min.z) aabb_min.z = pos.z;
-	  if (pos.z > aabb_max.z) aabb_max.z = pos.z;
-	}
+        Block block = get_block(data, {x, y, z});
+        if (block.material != Material::Air) {
+          glm::ivec3 pos = {x, y, z};
+          if (pos.x < aabb_min.x) aabb_min.x = pos.x;
+          if (pos.x > aabb_max.x) aabb_max.x = pos.x;
+          if (pos.y < aabb_min.y) aabb_min.y = pos.y;
+          if (pos.y > aabb_max.y) aabb_max.y = pos.y;
+          if (pos.z < aabb_min.z) aabb_min.z = pos.z;
+          if (pos.z > aabb_max.z) aabb_max.z = pos.z;
+        }
       }
     }
   }
